pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "anycode-py"
version = "0.4.0"
description = "Orchestrate autonomous AI agent teams with dependency-aware task scheduling, inter-agent messaging, and provider-agnostic LLM integration."
readme = "README.md"
license = "MIT"
requires-python = ">=3.12"
authors = [{ name = "Quantlix" }]
keywords = [
    "agent",
    "orchestration",
    "multi-agent",
    "llm",
    "ai",
    "tool-use",
    "team-collaboration",
    "task-scheduling",
    "anthropic",
    "openai",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries",
    "Typing :: Typed",
]
dependencies = [
    "pydantic>=2.0",
    "anthropic>=0.40",
    "openai>=1.50",
    "python-dotenv>=1.2.2",
]

[project.urls]
Homepage = "https://github.com/Quantlix/anycode"
Repository = "https://github.com/Quantlix/anycode"
Issues = "https://github.com/Quantlix/anycode/issues"
Changelog = "https://github.com/Quantlix/anycode/blob/main/anycode-python/CHANGELOG.md"

[project.optional-dependencies]
telemetry = [
    "opentelemetry-api>=1.20",
    "opentelemetry-exporter-otlp-proto-grpc>=1.20",
    "opentelemetry-sdk>=1.20",
]
persistence = [
    "aiosqlite>=0.20",
]
redis = [
    "redis[hiredis]>=5.0",
]
vector = [
    "chromadb>=0.5",
]
google = [
    "google-genai>=1.0",
]
ollama = [
    "httpx>=0.27",
]
bedrock = [
    "boto3>=1.34",
]
azure = [
    "openai>=1.50",
]
mcp = [
    "mcp>=1.0",
]

[dependency-groups]
dev = [
    "pyright>=1.1.408",
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "ruff>=0.8",
    "aiosqlite>=0.20",
    "redis[hiredis]>=5.0",
    "chromadb>=0.5",
    "google-genai>=1.0",
    "boto3>=1.34",
    "mcp>=1.0",
    "httpx>=0.27",
]

[tool.hatch.build.targets.wheel]
packages = ["src/anycode"]

[tool.ruff]
target-version = "py312"
line-length = 150

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-m 'not integration'"
markers = [
    "integration: requires Docker test infrastructure",
    "redis: requires Redis container",
    "postgres: requires PostgreSQL container",
    "chromadb: requires ChromaDB container",
]

[tool.pyright]
pythonVersion = "3.12"
typeCheckingMode = "standard"
exclude = ["docs/"]
